make_standalone_toolchain.py: support riscv64.
It makes little sense to have explicit code to _not_ support riscv64
that's about the same number of lines (and more of a special case) than
what we'd need to just support it. Removing
make_standalone_toolchain.py entirely would also be consistent, but the
current situation makes little sense. The only reason I'm looking at
this is because someone was already confused, and if we (rightly) don't
think we can fix everyone's existing build setups, I don't think we
should assume we can fix all the outdated documentation on the
internet...

Bug: https://github.com/android/ndk/issues/1990
Test: treehugger
Change-Id: Ie1b538559780fa9e0a4959aacb0af57e8bc1b433