Improve standalone toolchain API too low errors.
Previously, if the user didn't specify an API level, the standalone
toolchain would default to the minimum API level supported by the NDK.
If they did specify an API level, their input was also checked against
the NDK's minimum.

That wasn't right when the LP32 and LP64 minimum supported versions were
different, but became asymptomatic once the NDK's minimum became 21. Now
that we have rv64 with a minimum version greater than the NDK's minimum,
the problem is back again.

Extend the meta/abis.json data with the minimum API level for each ABI,
and check that in make_standalone_toolchain.py rather than the ABI
generic one in platforms.json.

This commit doesn't address the TODO literally. The data fed into
abis.json is still hard coded rather than scanned from the sysroot, but
on second thought I think I prefer that, because if the platform were to
remove an API level from the sysroot unexpectedly, we wouldn't want to
auto-pull-up the NDK to match, we'd want to see errors.

Bug: https://github.com/android/ndk/issues/1990
Test: added test
Change-Id: I29687372dafd557d019fef792a680a25bef5dcde