Preserve ANDROID_USE_LEGACY_TOOLCHAIN_FILE.
From the analysis in the bug, we suspect this is asymptomatic now that
we're back to defaulting this to off. The bug was only present in the
legacy toolchain, and since we default to that, forgetting the value and
getting the default doesn't actually change the value. If/when we do
switch back to using the new toolchain by default, this would avoid a
bug (the bug was originally discovered in r24, where it was the default
for a sufficiently new CMake).

Bug: https://github.com/android/ndk/issues/1921
Test: Manual, since there's no symptom in the current configuration
Change-Id: I58887afd9a325b5f756f8e17a4ff5088f2c83b10