Migrate ndk-stack to Path.
This makes it slightly easier to migrate the trace handling from str to
bytes because the Path APIs explicitly don't allow bytes inputs at all,
whereas os.path just doesn't allow mixing str and bytes. It's not nearly
as useful as I thought it would be when I started, but it's a useful
modernization in its own right, so might as well emrge it.

Bug: None
Test: ./checkbuild.py pythonlint
Change-Id: Ib378cabde6f72340fa14b74a65d9e2431042791a