Move ndkstack.find_* tests off mocks.
These are really just testing the mocks. Build a stub NDK in a temp dir
so we can test them more properly.

I did delete one test here. There was a separate test for finding the
linux-arm readelf instead of the linux-x86 readelf. That test wasn't
really testing that, but rather testing that the third argument (the
host tag) was used. It's not really a useful test, so I removed it.

Bug: None
Test: pytest tests/pytest
Change-Id: Ief33b6a42337741fc0ce6a3245ad8496d3fff9fb