Fix NDK bin path in ndk-gdb.
A similar problem was fixed in ndk-stack in r26b, but we forgot to check
the other things that had been converted to Python zipapps. This is the
only other one.

There's no easy test to add here. ndk-gdb appears to (understandably)
fail on all the paths that do not successfully connect the debugger. The
only exception is `--help`, but that also would exit before we'd reached
this error anyway. I'm not going to delay sending out this fix while we
wait for the chance to write some real ndk-gdb tests, so this will have
to get by with just a manual test.

Bug: https://github.com/android/ndk/issues/1994
Test: ran ndk-lldb, didn't see "could not find make" error
Change-Id: Ia11d7411926d5a24a415cb48c9e92572ee63626a