Add test for PAGE_SIZE presence.
PAGE_SIZE will not be consistent across all devices in the future:
https://source.android.com/docs/core/architecture/16kb-page-size/16kb.
As such, the `PAGE_SIZE` macro can't be reliably used at some point in
the future, and will eventually be removed from the header. For now,
test that it still exists so that we have a test to invert once we make
that change.

Bug: http://b/312546062
Test: ./run_tests.py --rebuild
Change-Id: I6bf46b7ca7def2f2e7a988fc26ed33009a61bb3e