Remove permanent known issue from the changelog.
Callign `dlclose` is and always will be a bad idea. This is already
documented in the API docs for `dlclose`, so we don't need to keep this
in the change log forever. It's not a change, is a platform issue rather
than an NDK issue, and the platform bug was fixed many years ago.

Bug: None
Test: None
Change-Id: I3a1f3269bc0f9b8bf70246505a65e69177634565